Sok vicces üres osztály

+++ src/team_qgli/Cape.java
@@ -1,7 +1,7 @@
 package team_qgli;
 
 public class Cape extends Equipment{
-    private float defPerc;
+    private long defPerc;
     @Override
     public void Collect(Virologist v) {
